+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 - Refactored the way to calculate the text styles
 - Added `game.config.stage` to modify the default stage
 - Moved to ESBuild and reduced the bundle size/build time
+- Added `game.config.defaultMenuChoiceColor` to set the default color for the menu choices
+- Hide the menu dialog when the menu has no prompt
 
 ## [0.4.0] - 2025/4/22